requirements.txt
dj-database-url
Django
pytz
python-decouple
jsonfield
djangorestframework
markdown
django-filter
coreapi
pygments
channels
channels_redis
pygal
pyparallel
paho-mqtt
pyFirmata
iso8601